requirements.txt
streamlit
groq
python-dotenv
httpx==0.27.0